+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Make the brackets print in the optional package install instructions
 
 ## 1.6.0
 * PDF previews in manual sorting windows